requirements.txt
# Core Dependencies for PTIP (Personal Trading Intelligence Platform)

# Web Framework
streamlit==1.50.0

# Data Analysis
pandas==2.3.3
numpy==2.3.3

# Trading API
fyers_apiv3==3.1.7

# Technical Analysis
ta==0.11.0

# Environment Variables
python-dotenv==1.1.1

# Database (SQLite is built-in with Python, no package needed)

# Additional utilities that may be needed
requests==2.31.0

